requirements.txt
absl-py==1.4.0
aiosignal==1.3.1
appdirs==1.4.4
attrs==23.1.0
casadi==3.5.5
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.4
-e .
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
distlib==0.3.7
dm-tree==0.1.8
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
filelock==3.12.4
fonttools==4.42.1
frozenlist==1.4.0
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.32
grpcio==1.51.3
Gymnasium
gymnasium-notices==0.0.1
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lazy-loader==0.3
lit==16.0.6
lz4==4.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
pettingzoo==1.24.0
Pillow==10.0.0
pkgutil-resolve-name==1.3.10
platformdirs==3.10.0
protobuf==4.24.3
psutil==5.9.5
pyarrow==13.0.0
pygame==2.5.1
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
ray==2.5.0
referencing==0.30.2
requests==2.31.0
rich
rpds-py==0.10.3
scikit-image==0.21.0
scipy==1.10.1
sentry-sdk==1.29.2
setproctitle==1.3.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
SuperSuit==3.9.0
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
tensorflow-probability==0.21.0
tifffile==2023.7.10
tinyscaler==1.2.6
torch==2.0.1
triton==2.0.0
typer==0.9.0
typing-extensions
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.21.0
wandb==0.15.8
zipp==3.16.2
